Added ability to import and export music songs in SongManager.
Manually editing all the songs one at a time is tedious; particularly if
you need to do it again for some reason (such as a new release of PM).
This functionality allows a user to export their current songs to a
directory with the *.brstm files named with the volume and title (as
well as the original game filename). The user can then replace file
contents, or change the title or volume of various songs before they
import the songs back (as long as the original game filename is in-tact).
This also allows easy sharing of a music 'build' as the exported songs can
be archived, shared and then imported.